+++ src/components/todo/TodoList.tsx
@@ -1,24 +1,11 @@
 import { useNavigate } from 'react-router-dom'
 import { useContext, useEffect } from 'react'
-import { getTodoList } from 'api/todo'
 import { ITodoData } from 'interfaces/ITodo'
 import TodoItem from './TodoItem'
-import useTodoContext from 'hooks/useTodoContext'
+import useGetTodoList from 'hooks/useGetTodoList'
 
 export default function TodoList() {
-  const { contextTodoList, setContextTodoList } = useTodoContext()
-
-  const navigate = useNavigate()
-
-  useEffect(() => {
-    getTodoList().then((data) => {
-      if (data === 401) {
-        navigate('/signin', { replace: true })
-        return
-      }
-      setContextTodoList?.(data)
-    })
-  }, [])
+  const contextTodoList = useGetTodoList()
 
   return (
     <>
